Added .env.example file as an example of environment variable configuration and renamed db_config file to db-config
