+++ .env.exmple
@@ -1,0 +1,15 @@
+#To run this project, you will need to add the following environment variables to your .env file
+
+DB_CONNECTION= tu conexion / your conenction
+
+DB_HOST= tu puerto de localhost / your port of localhost
+
+DB_USERNAME= tu nombre de usuario / your username
+
+DB_PASSWORD= tu contraseña / your password
+
+DB_NAME= tu base de datos o colecion / your data base or colection
+
+PORT= tu numero de puerto ejemplo 3000 / your port number for example 3000
+
+ENVIRONMENT = production
